fix(create): address review feedback on #333
- Co-locate interpolate unit coverage with the scaffold source.
