+++ packages/create/tests/scaffold.test.ts
@@ -3,7 +3,7 @@ import { existsSync, mkdirSync, readFileSync, rmSync, writeFileSync } from "node
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 
-import { interpolate, scaffold } from "../src/scaffold.ts";
+import { scaffold } from "../src/scaffold.ts";
 
 // ────────────────────────────────────────────────────────────────────────────
 // Test Helpers
@@ -57,46 +57,6 @@ function readOutputFile(relativePath: string): string {
 // ────────────────────────────────────────────────────────────────────────────
 // Tests
 // ────────────────────────────────────────────────────────────────────────────
-
-describe("interpolate", () => {
-	it("replaces multiple different variables", () => {
-		const result = interpolate("{{greeting}}, {{name}}!", {
-			greeting: "Hi",
-			name: "Crust",
-		});
-		expect(result).toBe("Hi, Crust!");
-	});
-
-	it("replaces repeated occurrences of the same variable", () => {
-		expect(interpolate("{{a}} and {{a}}", { a: "x" })).toBe("x and x");
-	});
-
-	it("leaves missing variables untouched", () => {
-		expect(interpolate("{{known}} and {{unknown}}", { known: "yes" })).toBe("yes and {{unknown}}");
-	});
-
-	it("returns original content when there are no placeholders", () => {
-		expect(interpolate("Hello, world!", { name: "unused" })).toBe("Hello, world!");
-	});
-
-	it("handles whitespace inside braces", () => {
-		expect(interpolate("{{ name }} and {{  spaced  }}", { name: "a", spaced: "b" })).toBe(
-			"a and b",
-		);
-	});
-
-	it("does not replace partial or malformed placeholders", () => {
-		expect(interpolate("{name} and {{}} and {{}}", { name: "x" })).toBe("{name} and {{}} and {{}}");
-	});
-
-	it("replaces with empty string values", () => {
-		expect(interpolate("before{{name}}after", { name: "" })).toBe("beforeafter");
-	});
-
-	it("handles underscores in variable names", () => {
-		expect(interpolate("{{my_var}}", { my_var: "value" })).toBe("value");
-	});
-});
 
 describe("scaffold", () => {
 	it("scaffolds a simple template with text files and interpolation", async () => {
